Remove unnecessary explicit cast

+++ src/hands/KM_HandStats.pas
@@ -497,7 +497,7 @@ var
 begin
   Result := 0;
   for UT := WARRIOR_MIN to WARRIOR_MAX do
-    Result := Result + Single(GetUnitQty(UT)) * WARRIORS_POWER_RATES[UT];
+    Result := Result + GetUnitQty(UT) * WARRIORS_POWER_RATES[UT];
 end;
